fix(ui): 개인 재화 소모 후 HUD 갱신 오류 수정

+++ Source/NeoSanctum/UI/HUD/NSOutRunGoodsWidget.h
@@ -24,7 +24,7 @@ public:
 
 	//영구재화 표시
 	UFUNCTION(BlueprintCallable, Category = "UI|Goods")
-	void SetCommonGoodsAmount(int32 NewAmount);
+	void SetCommonGoodsAmount(int64 NewAmount);
 
 private:
 	//아웃런 영구재화 텍스트
